FEM: unit tests, move solver framework and ccxtools test classes in separate modules
