+++ src/Mod/Fem/CMakeLists.txt
@@ -115,7 +115,9 @@ SET(FemSolverFenics_SRCS
 
 SET(FemTests_SRCS
     femtest/__init__.py
+    femtest/testccxtools.py
     femtest/testfemcommon.py
+    femtest/testsolverframework.py
     femtest/testtools.py
 )
