+++ src/Mod/Fem/TestFem.py
@@ -25,6 +25,8 @@
 # ***************************************************************************/
 
 
+# the order should be as follows:
+# common-, object-, mesh-, inout-, ccxtools-, solverframworktests
 from femtest.testfemcommon import FemTest
-from femtest.testfemcommon import FemCcxAnalysisTest
-from femtest.testfemcommon import SolverFrameWorkTest
+from femtest.testccxtools import FemCcxAnalysisTest
+from femtest.testsolverframework import SolverFrameWorkTest
